* audio/Midi.cc, audio/midi_drivers/* : Changed the way how MT32 Sysex data and timbre loading is handled to improve startup time. Things are now setup to preload the required timbers and patches used by songs as they are loaded, rather than loading then entire timbre bank as done previously. Added two config settings to control timbre precaching, "config/audio/midi/precacheTimbers/onStartup" (default no) and "config/audio/midi/precacheTimbers/onPlay" (default yes). It should be unnecessary to change those two settings. * other files : Fixing up various compiler warnings

+++ objs/chunks.cc
@@ -107,7 +107,7 @@ inline void Set_blocked_tile
 	{
 	uint16& val = blocked[ty*c_tiles_per_chunk + tx];
 					// Get mask for the bit0's:
-	uint16 mask0 = tmasks[ztiles]<<2*lift;
+	uint16 mask0 = (uint16)(tmasks[ztiles]<<2*lift);
 	uint16 mask1 = mask0<<1;	// Mask for the bit1's.
 	uint16 val0s = val&mask0;
 	uint16 Nval0s = (~val)&mask0;
@@ -134,7 +134,7 @@ inline void Clear_blocked_tile
 	{
 	uint16& val = blocked[ty*c_tiles_per_chunk + tx];
 					// Get mask for the bit0's:
-	uint16 mask0 = tmasks[ztiles]<<2*lift;
+	uint16 mask0 = (uint16)(tmasks[ztiles]<<2*lift);
 	uint16 mask1 = mask0<<1;	// Mask for the bit1's.
 	uint16 val0s = val&mask0;
 	uint16 Nval0s = (~val)&mask0;
